Make all http requests explictly ipv4 when using 8.6.
There is currently a bug in the ipv6 socket handling in 8.6 where if a
site has an ipv6 address and does not provide a services on that
address then socket -async will return an error to tcl. A synchronous
socket will try ipv6 and then ipv4 before returning to tcl. To work
around this we can force http requests to use ipv4 until this is fixed
so that async sockets operate like the synchronous ones.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>